requirements.txt
Flask==1.1.1
InstagramAPI==1.0.2
requests==2.22.0
pandas==0.23.4
pathlib2==2.3.2
google-cloud-storage==1.13.0
openpyxl==2.5.11